Fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45400 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/userport/userport_wic64.c
@@ -2222,7 +2222,7 @@ static void do_command(void)
             send_reply_revised(CLIENT_ERROR, "", NULL, 0, NULL);
         break;
     case WIC64_CMD_REBOOT:
-        wic64_sleep_cycles(3 * machine_get_cycles_per_second()); /* emulated a 3s reboot */
+        wic64_sleep_cycles((int)(3 * machine_get_cycles_per_second())); /* emulated a 3s reboot */
         userport_wic64_reset();
         break;
     case WIC64_CMD_SET_TRANSFER_TIMEOUT:
